routes of pages add to flask
